Add files via upload
Implemented both encryption and decryption of message using Elliptical Curve Cryptography on the Zynq-7000 FPGA. Verilog code was synthesized and implemented on Xilinx Vivado. The generated bitstream was used to configure the FPGA. Input message was given through DIP switches and output was verified using LED's present on the hardware.